Update WinterLand
Add 2 green candyzombie + 1 black candyzombie image
