+++ Assets/Animations/XmasAnimation/XmasGreenZombieAnimation.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: c29fb8bce80b2c747a0b6877a6160e1f
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
