+++ Assets/Animations/XmasAnimation/XmasGreenZombieAnimation/Zombie_attack.anim
@@ -1,0 +1,92 @@
+%YAML 1.1
+%TAG !u! tag:unity3d.com,2011:
+--- !u!74 &7400000
+AnimationClip:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_Name: Zombie_attack
+  serializedVersion: 6
+  m_Legacy: 0
+  m_Compressed: 0
+  m_UseHighQualityCurve: 1
+  m_RotationCurves: []
+  m_CompressedRotationCurves: []
+  m_EulerCurves: []
+  m_PositionCurves: []
+  m_ScaleCurves: []
+  m_FloatCurves: []
+  m_PPtrCurves:
+  - curve:
+    - time: 0
+      value: {fileID: 21300096, guid: 0f9b2339c6e91924a966dee77b695667, type: 3}
+    - time: 0.125
+      value: {fileID: 21300100, guid: 0f9b2339c6e91924a966dee77b695667, type: 3}
+    - time: 0.25
+      value: {fileID: 21300102, guid: 0f9b2339c6e91924a966dee77b695667, type: 3}
+    - time: 0.375
+      value: {fileID: 21300104, guid: 0f9b2339c6e91924a966dee77b695667, type: 3}
+    - time: 0.5
+      value: {fileID: 21300106, guid: 0f9b2339c6e91924a966dee77b695667, type: 3}
+    - time: 0.625
+      value: {fileID: 21300108, guid: 0f9b2339c6e91924a966dee77b695667, type: 3}
+    - time: 0.75
+      value: {fileID: 21300110, guid: 0f9b2339c6e91924a966dee77b695667, type: 3}
+    - time: 0.875
+      value: {fileID: 21300112, guid: 0f9b2339c6e91924a966dee77b695667, type: 3}
+    - time: 1
+      value: {fileID: 21300114, guid: 0f9b2339c6e91924a966dee77b695667, type: 3}
+    attribute: m_Sprite
+    path: 
+    classID: 212
+    script: {fileID: 0}
+  m_SampleRate: 8
+  m_WrapMode: 0
+  m_Bounds:
+    m_Center: {x: 0, y: 0, z: 0}
+    m_Extent: {x: 0, y: 0, z: 0}
+  m_ClipBindingConstant:
+    genericBindings:
+    - serializedVersion: 2
+      path: 0
+      attribute: 0
+      script: {fileID: 0}
+      typeID: 212
+      customType: 23
+      isPPtrCurve: 1
+    pptrCurveMapping:
+    - {fileID: 21300096, guid: 0f9b2339c6e91924a966dee77b695667, type: 3}
+    - {fileID: 21300100, guid: 0f9b2339c6e91924a966dee77b695667, type: 3}
+    - {fileID: 21300102, guid: 0f9b2339c6e91924a966dee77b695667, type: 3}
+    - {fileID: 21300104, guid: 0f9b2339c6e91924a966dee77b695667, type: 3}
+    - {fileID: 21300106, guid: 0f9b2339c6e91924a966dee77b695667, type: 3}
+    - {fileID: 21300108, guid: 0f9b2339c6e91924a966dee77b695667, type: 3}
+    - {fileID: 21300110, guid: 0f9b2339c6e91924a966dee77b695667, type: 3}
+    - {fileID: 21300112, guid: 0f9b2339c6e91924a966dee77b695667, type: 3}
+    - {fileID: 21300114, guid: 0f9b2339c6e91924a966dee77b695667, type: 3}
+  m_AnimationClipSettings:
+    serializedVersion: 2
+    m_AdditiveReferencePoseClip: {fileID: 0}
+    m_AdditiveReferencePoseTime: 0
+    m_StartTime: 0
+    m_StopTime: 1.125
+    m_OrientationOffsetY: 0
+    m_Level: 0
+    m_CycleOffset: 0
+    m_HasAdditiveReferencePose: 0
+    m_LoopTime: 1
+    m_LoopBlend: 0
+    m_LoopBlendOrientation: 0
+    m_LoopBlendPositionY: 0
+    m_LoopBlendPositionXZ: 0
+    m_KeepOriginalOrientation: 0
+    m_KeepOriginalPositionY: 1
+    m_KeepOriginalPositionXZ: 0
+    m_HeightFromFeet: 0
+    m_Mirror: 0
+  m_EditorCurves: []
+  m_EulerEditorCurves: []
+  m_HasGenericRootTransform: 0
+  m_HasMotionFloatCurves: 0
+  m_Events: []
