+++ Assets/Animations/XmasAnimation/XmasGreenZombieAnimation/Zombie_attack.anim.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: e6bed13668ef7264db1f898585fb8908
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 7400000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
