+++ Assets/Animations/XmasAnimation/XmasGreenZombieAnimation/Zombie_die.anim.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: dde93706116955c49b218efb345e88f3
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 7400000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
