+++ Assets/Animations/XmasAnimation/XmasGreenZombieAnimation/Zombie_walk.anim.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 1f51744e6500c034a92ca80fbcf3c9d9
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 7400000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
