+++ Assets/Animations/XmasAnimation/XmasGreenZombieAnimation/xmas zombie sprite(1)_0.controller.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 5e24c5d4d1897f9439dcfc7bcb4b62b4
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 9100000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
