+++ Assets/Scripts/Enemy Scripts/XmasEnemiesScripts/GreenZMovement.cs
@@ -1,0 +1,47 @@
+﻿using System.Collections;
+using System.Collections.Generic;
+using UnityEngine;
+
+public class GreenZMovement : MonoBehaviour
+{
+    public float walkSpeed = 1.2f;
+    public bool walkLeft;
+
+
+    void Start()
+    {
+        StartCoroutine(changeDirection());
+    }
+
+
+    void Update()
+    {
+        enemyWalk();
+    }
+
+    private void enemyWalk()
+    {
+        Vector3 temp = transform.position;
+        Vector3 tempScale = transform.localScale;
+        if (walkLeft)
+        {
+            temp.x -= walkSpeed * Time.deltaTime;
+            tempScale.x = -Mathf.Abs(tempScale.x);
+        }
+        else
+        {
+            temp.x += walkSpeed * Time.deltaTime;
+            tempScale.x = Mathf.Abs(tempScale.x);
+        }
+
+        transform.position = temp;
+        transform.localScale = tempScale;
+    }
+
+    IEnumerator changeDirection()
+    {
+        yield return new WaitForSeconds(14f);    // Attende un tempo pari a 3 secondi.
+        walkLeft = !walkLeft;   // Cambia direzione.
+        StartCoroutine(changeDirection());
+    }
+}
