+++ Assets/Scripts/Enemy Scripts/XmasEnemiesScripts/GreenZMovement.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: e73e2a36b3eba944e9ec8c9b15409af9
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
