+++ Assets/Scripts/Enemy Scripts/XmasEnemiesScripts/GreenZombieMovement.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: ad7efa7f384952849a76c514bf63de14
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
