+++ Assets/Scripts/Enemy Scripts/XmasEnemiesScripts/SnowballMovement.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: e5854a5f053fedf488c57408b4b3ff91
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
