+++ Assets/Scripts/Enemy Scripts/XmasEnemiesScripts/YellowZombieMovement.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 0dfe18c10cb2e4a48841b388b5ea06a3
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
