+++ Assets/Scripts/Enemy Scripts/XmasEnemiesScripts/YetiWhiteMovement.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 410fddc4d2bbf4b4690e6d843e8833fe
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
