--- Assets/Scripts/Enemy Scripts/YetiBlackMovement.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: cf7ac990ec18ebd43b349102abf03f6f
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
